+++ fakeredis/__init__.py
@@ -1,5 +1,6 @@
 import sys
 
+from . import typing
 from ._connection import (
     FakeRedis,
     FakeStrictRedis,
@@ -20,12 +21,7 @@ else:
             raise NotImplementedError("TcpFakeServer is only available in Python 3.11+")
 
 
-try:
-    from importlib import metadata
-except ImportError:  # for Python < 3.8
-    import importlib_metadata as metadata  # type: ignore
-
-__version__ = metadata.version("fakeredis")
+__version__ = typing.lib_version
 __author__ = "Daniel Moran"
 __maintainer__ = "Daniel Moran"
 __email__ = "daniel@moransoftware.ca"