+++ fakeredis/_valkey.py
@@ -11,6 +11,7 @@ def _validate_server_type(args_dict: Dict[str, Any]) -> None:
     if "server_type" in args_dict and args_dict["server_type"] != "valkey":
         raise ValueError("server_type must be valkey")
     args_dict.setdefault("server_type", "valkey")
+    args_dict.setdefault("client_class", valkey.Valkey)
 
 
 class FakeValkey(FakeRedisMixin, valkey.Valkey):
@@ -20,7 +21,6 @@ class FakeValkey(FakeRedisMixin, valkey.Valkey):
 
     @classmethod
     def from_url(cls, *args: Any, **kwargs: Any) -> Self:
-        _validate_server_type(kwargs)
         return super().from_url(*args, **kwargs)
 
 
@@ -31,16 +31,15 @@ class FakeStrictValkey(FakeRedisMixin, valkey.StrictValkey):
 
     @classmethod
     def from_url(cls, *args: Any, **kwargs: Any) -> Self:
-        _validate_server_type(kwargs)
         return super().from_url(*args, **kwargs)
 
 
 class FakeAsyncValkey(FakeAsyncRedisMixin, valkey.asyncio.Valkey):
     def __init__(self, *args: Any, **kwargs: Any) -> None:
+        kwargs.setdefault("client_class", valkey.asyncio.Valkey)
         _validate_server_type(kwargs)
         super(FakeAsyncValkey, self).__init__(*args, **kwargs)
 
     @classmethod
     def from_url(cls, *args: Any, **kwargs: Any) -> Self:
-        _validate_server_type(kwargs)
         return super().from_url(*args, **kwargs)