+++ fakeredis/stack/_cf_mixin.py
@@ -47,34 +47,38 @@ class CFCommandsMixin:
 
     @staticmethod
     def _cf_exist(key: CommandItem, item: bytes) -> int:
-        return 1 if (item in key.value) else 0
+        return 1 if (key.value is not None and item in key.value) else 0
 
-    @command(name="CF.ADD", fixed=(Key(ScalableCuckooFilter), bytes), repeat=())
+    @command(name="CF.ADD", fixed=(Key(ScalableCuckooFilter), bytes), repeat=(), flags=msgs.FLAG_DO_NOT_CREATE)
     def cf_add(self, key: CommandItem, value: bytes) -> int:
         return CFCommandsMixin._cf_add(key, value)
 
-    @command(name="CF.ADDNX", fixed=(Key(ScalableCuckooFilter), bytes), repeat=())
+    @command(name="CF.ADDNX", fixed=(Key(ScalableCuckooFilter), bytes), repeat=(), flags=msgs.FLAG_DO_NOT_CREATE)
     def cf_addnx(self, key: CommandItem, value: bytes) -> int:
         if value in key.value:
             return 0
         return CFCommandsMixin._cf_add(key, value)
 
-    @command(name="CF.COUNT", fixed=(Key(ScalableCuckooFilter), bytes), repeat=())
+    @command(name="CF.COUNT", fixed=(Key(ScalableCuckooFilter), bytes), repeat=(), flags=msgs.FLAG_DO_NOT_CREATE)
     def cf_count(self, key: CommandItem, item: bytes) -> int:
-        return 1 if self._cf_exist(key, item) else 0  # todo
+        if key.value is None:
+            return 0
+        if type(key.value) is not ScalableCuckooFilter:
+            raise SimpleError(msgs.WRONGTYPE_MSG)
+        return key.value.count(item)
 
-    @command(name="CF.DEL", fixed=(Key(ScalableCuckooFilter), bytes), repeat=())
+    @command(name="CF.DEL", fixed=(Key(ScalableCuckooFilter), bytes), repeat=(), flags=msgs.FLAG_DO_NOT_CREATE)
     def cf_del(self, key: CommandItem, value: bytes) -> int:
         if key.value is None:
             raise SimpleError(msgs.NOT_FOUND_MSG)
         res = key.value.delete(value)
         return 1 if res else 0
 
-    @command(name="CF.EXISTS", fixed=(Key(ScalableCuckooFilter), bytes), repeat=())
+    @command(name="CF.EXISTS", fixed=(Key(ScalableCuckooFilter), bytes), repeat=(), flags=msgs.FLAG_DO_NOT_CREATE)
     def cf_exist(self, key: CommandItem, value: bytes) -> int:
         return CFCommandsMixin._cf_exist(key, value)
 
-    @command(name="CF.INFO", fixed=(Key(),), repeat=())
+    @command(name="CF.INFO", fixed=(Key(),), repeat=(), flags=msgs.FLAG_DO_NOT_CREATE)
     def cf_info(self, key: CommandItem) -> Dict[bytes, Any]:
         if key.value is None or type(key.value) is not ScalableCuckooFilter:
             raise SimpleError("...")