+++ docs/about/changelog.md
@@ -7,12 +7,43 @@ tags:
 toc_depth: 2
 ---
 
+## v2.31.0 - 2025-08-10
+
+### 🚀 Features
+
+- Add native cjson support for lua redis scripts #399 @ItsDrike
+
+### 🐛 Bug Fixes
+
+- version init argument typing #401
+- include `lib_name` and `lib_version` in `FakeBaseConnectionMixin`
+- when using `FakeValkey`, raise `valkey.ResponseError` instead of `redis.ResponseError` #402
+- fix issue with pytest-asyncio #404 @seifertm
+
+## v2.30.3 - 2025-08-01
+
+### 🐛 Bug Fixes
+
+- Import `Callable` from `typing` to support python 3.8 #398
+- FakeValkey init arguments #400
+
+## v2.30.2 - 2025-08-01
+
+### 🐛 Bug Fixes
+
+- Import `Callable` from `typing` to support python 3.8 #398
+
+### 🧰 Maintenance
+
+- Refactor tests
+- `FakeValkey` uses valkey-py instead of redis-py
+
 ## v2.30.1 - 2025-06-20
 
 ### 🐛 Bug Fixes
 
 - Move `LICENSE` file to fakeredis/ on build #395
-- Show warning for deprecated parameter only if parameter is included in instantiation #384
+- Show warning for deprecated parameter only if the parameter is included in instantiation #384
 
 ### 🧰 Maintenance
 
@@ -27,7 +58,7 @@ toc_depth: 2
 
 ### 🐛 Bug Fixes
 
-- aio.FakeRedis: Using `inspect` to get `__init__` paramaters #384
+- aio.FakeRedis: Using `inspect` to get `__init__` parameters #384
 - Discrepancy when using xread with resp3 #393
 
 ## v2.29.0 - 2025-05-01